pyproject.toml
[tool.poetry]
name = "LORIN"
version = "0.1.0"
description = "LORIN - Log Analysis and Retrieval Intelligence Network"
authors = ["BALab in CAU"]
packages = [{include = "LORIN"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langgraph = ">=0.6.0"
langsmith = ">=0.4.8"
langchain = ">=0.3.27,<0.4.0"
langchain-core = ">=0.3.27,<0.4.0"
langchain-openai = ">=0.3.28,<0.4.0"
langchain-anthropic = ">=0.3.18,<0.4.0"
langchain-google-genai = ">=2.1.8,<3.0.0"
langchain-community = ">=0.3.0,<0.4.0"
python-dotenv = ">=1.1.1,<2.0.0"
numpy = ">=2.3.2,<3.0.0"
faiss-cpu = ">=1.8.0"
pandas = ">=2.2.0,<3.0.0"
tqdm = ">=4.66.0,<5.0.0"
sentence-transformers = ">=2.5.1,<3.0.0"
langchain-together = ">=0.2.0"
together = ">=1.3.0"
langchain-huggingface = ">=0.3.1,<0.4.0"
transformers = ">=4.56.1,<5.0.0"
safetensors = ">=0.6.2,<0.7.0"
torch = {version = "2.6.0+cu124", source = "pytorch-cu124"}
hf-transfer = ">=0.1.9,<0.2.0"
rank-bm25 = ">=0.2.2,<0.3.0"
colbert-ai = ">=0.2,<0.3"
accelerate = ">=1.10.1,<2.0.0"
huggingface-hub = ">=0.35.0,<0.36.0"
einops = ">=0.8.1,<0.9.0"
hf-xet = ">=1.1.10,<2.0.0"
colorama = ">=0.4.6,<0.5.0"
streamlit = ">=1.28.0,<2.0.0"
matplotlib = ">=3.7.0,<4.0.0"
seaborn = ">=0.12.0,<0.14.0"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
